JWT Authentication, and backend Completion

+++ backend/controllers/goalController.js
@@ -1,11 +1,12 @@
 const asyncHandler = require('express-async-handler')
-
 const Goal = require('../models/goalModel')
+const User = require('../models/userModel')
+
 //@desc GEt Goals
 // @route GET/api/goals
 //access Private
 const getGoals = asyncHandler( async (req, res) => {
-    const goals = await Goal.find()
+    const goals = await Goal.find({ user: req.user.id})
 
     res.status(200).json (goals)
 })
@@ -18,7 +19,8 @@ const setGoals = asyncHandler( async (req, res) => {
      throw new Error('PLease addd a text field')
    }
     const goal = await Goal.create({
-       text: req.body.text 
+       text: req.body.text,
+       user: req.user.id, 
 })
     res.status(200).json (goal)
 })
@@ -31,6 +33,18 @@ const updateGoal = asyncHandler( async (req, res) => {
     if(!goal) {
         res.status(400)
         throw new Error('Goal not found')
+    }
+    const user = await User.findById(req.user.id)
+    // check for use
+    if(!user) {
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    //Make sure logged in user mathces the goal user
+    if(goal.user.toString() !== user.id){
+        res.status(401)
+        throw new Error('User not authorized')
     }
 
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
@@ -48,6 +62,19 @@ const deleteGoal = asyncHandler( async (req, res) => {
     if(!goal) {
         res.status(404)
         throw new Error('Goal not found')
+    }
+
+    const user = await User.findById(req.user.id)
+    // check for user
+    if(!user) {
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    //Make sure logged in user mathces the goal user
+    if(goal.user.toString() !== user.id){
+        res.status(401)
+        throw new Error('User not authorized')
     }
 
     await goal.deleteOne()
